Implement an attribute cache
While parsing tags, we need to drop duplicate attribute. The previous
implementation is an O(n^2) algorithm. For the common case of just a few
attributes, this is very fast. With a very large number of attributes,
this becomes untenable.

rb_tree.[hc] implements a red-black tree specialized to using strings as
keys and no values. The implementation supports inserting strings and
freeing the whole tree. In particular, there is no support for deletion.